user service refacto
