+++ apps/shelve/server/api/user/index.delete.ts
@@ -1,8 +1,6 @@
-import { UserService } from '~~/server/services/user.service'
-
 export default eventHandler(async (event) => {
   const { user } = await requireUserSession(event)
-  await new UserService().deleteUserById(user.id)
+  await useDrizzle().delete(tables.users).where(eq(tables.users.id, user.id))
   return {
     statusCode: 200,
     message: 'User deleted',
