+++ apps/shelve/server/routes/auth/google.ts
@@ -1,5 +1,5 @@
 import { AuthType } from '@shelve/types'
-import { UserService } from '~~/server/services/user.service'
+import { handleOAuthUser } from '~~/server/services/user'
 import { userSchema } from '~~/server/database/zod'
 
 export default defineOAuthGoogleEventHandler({
@@ -8,7 +8,7 @@ export default defineOAuthGoogleEventHandler({
   },
   async onSuccess(event, { user, tokens }) {
     try {
-      const _user = await new UserService().handleOAuthUser({
+      const _user = await handleOAuthUser({
         email: user.email,
         avatar: user.picture,
         username: `${user.given_name}_${user.family_name}`,
